requirements-extra.txt
# FORMAT
# Put your extra requirements here in the following format
#
# package[version_required] # tag1, tag2, ...
#
# Allow at least one space between the package name/version and '#'
#
# Only one `#` is allowed per line. Lines starting with `#` are ignored.

flake8 #                       dev
sphinx #                       dev, doc
ipython #                      dev
numpydoc #                     dev, doc
sphinx-copybutton #            dev, doc
sphinx-panels #                dev, doc
pydata-sphinx-theme #          dev, doc
myst-parser==0.15.2 #          dev, doc
myst-nb==0.13.2 #              dev, doc
